+++ packages/pluggableWidgets/carousel-web/src/Carousel.tsx
@@ -1,16 +1,18 @@
 import { createElement, ReactNode, useEffect, useState, useCallback, ReactElement } from "react";
+import { ValueStatus } from "mendix";
 import { executeAction } from "@mendix/piw-utils-internal";
 import { CarouselContainerProps } from "../typings/CarouselProps";
 import { Carousel as CarouselComponent } from "./components/Carousel";
 import loadingCircleSvg from "./ui/loading-circle.svg";
 import "./ui/Carousel.scss";
+import classNames from "classnames";
 
 export function Carousel(props: CarouselContainerProps): ReactNode {
     const { dataSource, content, showPagination, loop, tabIndex, navigation, animation, delay, autoplay } = props;
     const [loading, setLoading] = useState(true);
     const onClick = useCallback(() => executeAction(props.onClickAction), [props.onClickAction]);
     useEffect(() => {
-        if (dataSource?.status === "available") {
+        if (dataSource?.status === ValueStatus.Available) {
             setLoading(false);
         }
     }, [dataSource]);
@@ -19,9 +21,10 @@ export function Carousel(props: CarouselContainerProps): ReactNode {
     const renderCarousel = (): ReactElement => {
         return (
             <CarouselComponent
+                className={props.class}
+                tabIndex={tabIndex}
                 pagination={showPagination}
                 loop={loop}
-                tabIndex={tabIndex}
                 animation={animation}
                 autoplay={autoplay}
                 delay={delay}
@@ -33,7 +36,11 @@ export function Carousel(props: CarouselContainerProps): ReactNode {
         );
     };
     const renderLoading = (): ReactNode => {
-        return <img src={loadingCircleSvg} className="widget-carousel-loading-spinner" alt="" aria-hidden />;
+        return (
+            <div className={classNames(props.class, "widget-carousel")} tabIndex={tabIndex}>
+                <img src={loadingCircleSvg} className="widget-carousel-loading-spinner" alt="" aria-hidden />
+            </div>
+        );
     };
     return loading ? renderLoading() : renderCarousel();
 }