+++ packages/pluggableWidgets/carousel-web/src/components/Carousel.tsx
@@ -1,23 +1,37 @@
 import { createElement, ReactNode, ReactElement } from "react";
 import { ObjectItem } from "mendix";
+import classNames from "classnames";
 import Swiper, { SwiperOptions, Navigation, Pagination, EffectFade, Autoplay } from "swiper";
 import { Swiper as SwiperReact, SwiperSlide } from "swiper/react";
 
-export interface CarouselProps<T extends ObjectItem> {
+export interface CarouselProps {
     pagination: boolean;
     loop: boolean;
-    tabIndex?: number | undefined;
     animation?: boolean;
     autoplay?: boolean;
     delay?: number;
     navigation: boolean;
+    className: string;
+    tabIndex?: number | undefined;
     onClick: () => void;
-    items: T[];
-    itemRenderer: (renderWrapper: (children: ReactNode) => ReactElement, item: T) => ReactNode;
+    items: ObjectItem[];
+    itemRenderer: (renderWrapper: (children: ReactNode) => ReactElement, item: ObjectItem) => ReactNode;
 }
 
-export function Carousel<T extends ObjectItem>(props: CarouselProps<T>): ReactElement {
-    const { items, pagination, loop, tabIndex, animation, autoplay, delay, navigation, itemRenderer, onClick } = props;
+export function Carousel(props: CarouselProps): ReactElement {
+    const {
+        items,
+        pagination,
+        loop,
+        animation,
+        autoplay,
+        delay,
+        navigation,
+        className,
+        tabIndex,
+        itemRenderer,
+        onClick
+    } = props;
     Swiper.use([Navigation, Pagination, EffectFade, Autoplay]);
     const options: SwiperOptions = {
         slidesPerView: 1,
@@ -32,11 +46,10 @@ export function Carousel<T extends ObjectItem>(props: CarouselProps<T>): ReactEl
         })
     };
     return (
-        <div className={"widget-carousel"} tabIndex={tabIndex}>
+        <div className={classNames(className, "widget-carousel")} tabIndex={tabIndex}>
             <SwiperReact {...options} onClick={onClick}>
-                {items.length > 0 &&
-                    itemRenderer &&
-                    items?.map((item: T, key: number) =>
+                {itemRenderer &&
+                    items?.map((item: ObjectItem, key: number) =>
                         itemRenderer((children: ReactNode) => <SwiperSlide key={key}>{children}</SwiperSlide>, item)
                     )}
             </SwiperReact>